Move translation viewing to admin namespace
