+++ app/controllers/admin_controller.rb
@@ -1,0 +1,3 @@
+class AdminController < ApplicationController
+  before_action :authenticate_admin!
+end
