+++ app/helpers/admin/dashboard_helper.rb
@@ -1,0 +1,2 @@
+module Admin::DashboardHelper
+end
