+++ app/helpers/admin/translations_helper.rb
@@ -1,0 +1,2 @@
+module Admin::TranslationsHelper
+end
