+++ app/views/admin/dashboard/index.html.erb
@@ -1,0 +1,2 @@
+<h1><%= t('.administration') %></h1>
+<%= link_to t('.view_translations'), admin_translations_path %>
