+++ app/views/admin/translations/_translation_pagination.html.erb
