+++ app/views/admin/translations/index.json.jbuilder
