+++ app/views/admin/translations/show.html.erb
