+++ app/views/admin/translations/show.json.jbuilder
