+++ config/locales/views/admin/de.yml
@@ -1,0 +1,6 @@
+de:
+  admin:
+    dashboard:
+      index:
+        administration: "Administration"
+        view_translations: "Übersetzungen anschauen"
