+++ test/controllers/admin/dashboard_controller_test.rb
@@ -1,0 +1,15 @@
+require 'test_helper'
+
+class Admin::DashboardControllerTest < ActionController::TestCase
+  include Devise::TestHelpers
+
+  setup do
+    login
+  end
+
+  test "should get index" do
+    get :index
+    assert_response :success
+  end
+
+end
