+++ test/test_helper.rb
@@ -10,4 +10,8 @@ class ActiveSupport::TestCase
   fixtures :all
 
   # Add more helper methods to be used by all tests here...
+  def login
+    admin = admins(:admin)
+    sign_in admin
+  end
 end
